+++ changeLog.md
@@ -28,6 +28,14 @@
 
 ## 29-09-2023
 ### What's new
+- Setup axe playwright accessibility testing for all pages
+- Fixed any issues found by the axe tests
+- updated git actions to run the axe tests
+### Why was this change made?
+- To ensure that future changes don't break accessibility
+
+## 29-09-2023
+### What's new
 - decrease wales obscure opacity
 ### Why was this change made?
 - Some data we have is on wales, so we need to obscure it less